requirements.txt
undetected-chromedriver
selenium
python-dotenv
mysql-connector-python
redis
